unit test serlization error
