+++ code_generation/templates/power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/meta_gen/attribute_classes.hpp.jinja
@@ -13,6 +13,7 @@
 #include "../../power_grid_model.hpp" // NOLINT
 #include "../../three_phase_tensor.hpp" // NOLINT
 #include "../meta_data.hpp" // NOLINT
+#include "gen_getters.hpp" // NOLINT
 #include "../{{ name }}.hpp" // NOLINT
